+++ CHANGELOG.md
@@ -27,6 +27,13 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Breaking
 
 - Upgraded solana/agave dependencies to agave 3. solana-storage.proto file updates adds a new `message_hash` field
+## 2025-09-11
+
+- yellowstone-grpc-geyser-9.0.1
+
+### Fixes
+
+- Fixed unload in grpc geyser plugin.
 
 ## 2025-08-04